modify int to float in shaders for compile on mac

+++ neurolabi/gui/Resources/shader/cube_wboit.vert
@@ -1,7 +1,7 @@
 // "Weighted Blended Order-Independent Transparency" technique by Morgan McGuire and Louis Bavoil
 
 attribute vec3 vPosition;
-attribute int vNormal;
+attribute float vNormal;
 //attribute vec3 vNormal;
 //attribute vec4 vColor;
